+++ lib/app/app_dependencies.dart
@@ -4,17 +4,13 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import 'package:flutter_web_portfolio/app/controllers/cursor_controller.dart';
-import 'package:flutter_web_portfolio/app/controllers/personalization_controller.dart';
 import 'package:flutter_web_portfolio/app/controllers/scene_director.dart';
 import 'package:flutter_web_portfolio/app/controllers/scroll_controller.dart';
 import 'package:flutter_web_portfolio/app/controllers/sound_controller.dart';
 import 'package:flutter_web_portfolio/app/data/providers/assets_provider.dart';
-import 'package:flutter_web_portfolio/app/data/providers/github_provider.dart';
 import 'package:flutter_web_portfolio/app/data/providers/local_storage_provider.dart';
-import 'package:flutter_web_portfolio/app/data/providers/medium_provider.dart';
 import 'package:flutter_web_portfolio/app/data/repositories/language_repository_impl.dart';
 import 'package:flutter_web_portfolio/app/features/language/application/language_cubit.dart';
-import 'package:flutter_web_portfolio/app/services/visitor_analytics.dart';
 
 /// Explicit, application-owned dependency graph.
 ///
@@ -27,9 +23,6 @@ final class AppDependencies {
     required this.sceneDirector,
     required this.cursorController,
     required this.soundController,
-    required this.personalizationController,
-    required this.mediumProvider,
-    required this.githubProvider,
   });
 
   final LanguageCubit languageCubit;
@@ -37,9 +30,6 @@ final class AppDependencies {
   final SceneDirector sceneDirector;
   final CursorController cursorController;
   final SoundController soundController;
-  final PersonalizationController personalizationController;
-  final MediumProvider mediumProvider;
-  final GitHubProvider githubProvider;
 
   static Future<AppDependencies> bootstrap() async {
     final assetsProvider = AssetsProvider();
@@ -55,9 +45,6 @@ final class AppDependencies {
     final sceneDirector = SceneDirector(scrollController: scrollController);
     final cursorController = CursorController();
     final soundController = SoundController();
-    final personalizationController = PersonalizationController(
-      analytics: VisitorAnalytics(storageProvider: storageProvider),
-    );
 
     return AppDependencies._(
       languageCubit: languageCubit,
@@ -65,9 +52,6 @@ final class AppDependencies {
       sceneDirector: sceneDirector,
       cursorController: cursorController,
       soundController: soundController,
-      personalizationController: personalizationController,
-      mediumProvider: MediumProvider(),
-      githubProvider: GitHubProvider(),
     );
   }
 
@@ -76,7 +60,6 @@ final class AppDependencies {
     await scrollController.close();
     await cursorController.close();
     await soundController.close();
-    await personalizationController.close();
     await languageCubit.close();
   }
 }
@@ -104,37 +87,24 @@ final class _AppRuntimeState extends State<AppRuntime> {
   }
 
   @override
-  Widget build(BuildContext context) => MultiRepositoryProvider(
+  Widget build(BuildContext context) => MultiBlocProvider(
     providers: [
-      RepositoryProvider<MediumProvider>.value(
-        value: widget.dependencies.mediumProvider,
+      BlocProvider<LanguageCubit>.value(
+        value: widget.dependencies.languageCubit,
       ),
-      RepositoryProvider<GitHubProvider>.value(
-        value: widget.dependencies.githubProvider,
+      BlocProvider<AppScrollController>.value(
+        value: widget.dependencies.scrollController,
+      ),
+      BlocProvider<SceneDirector>.value(
+        value: widget.dependencies.sceneDirector,
+      ),
+      BlocProvider<CursorController>.value(
+        value: widget.dependencies.cursorController,
+      ),
+      BlocProvider<SoundController>.value(
+        value: widget.dependencies.soundController,
       ),
     ],
-    child: MultiBlocProvider(
-      providers: [
-        BlocProvider<LanguageCubit>.value(
-          value: widget.dependencies.languageCubit,
-        ),
-        BlocProvider<AppScrollController>.value(
-          value: widget.dependencies.scrollController,
-        ),
-        BlocProvider<SceneDirector>.value(
-          value: widget.dependencies.sceneDirector,
-        ),
-        BlocProvider<CursorController>.value(
-          value: widget.dependencies.cursorController,
-        ),
-        BlocProvider<SoundController>.value(
-          value: widget.dependencies.soundController,
-        ),
-        BlocProvider<PersonalizationController>.value(
-          value: widget.dependencies.personalizationController,
-        ),
-      ],
-      child: widget.child,
-    ),
+    child: widget.child,
   );
 }