+++ lib/app/modules/home/sections/contact_section.dart
@@ -1,1 +1,0 @@
-export 'contact/contact_section.dart';