+++ lib/app/modules/home/sections/projects_section.dart
@@ -1,1 +1,0 @@
-export 'projects/projects_section.dart';